+++ backend/mwe/models.py
@@ -1,19 +1,11 @@
 from django.db import models
 
-from rest_framework import serializers
-
 
 class CanonicalForm(models.Model):
     text = models.TextField(unique=True)
 
     def __str__(self):
         return self.text
-
-
-class CanonicalFormSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = CanonicalForm
-        fields = ['id', 'text']
 
 
 class XPathQuery(models.Model):
@@ -25,15 +17,3 @@ class XPathQuery(models.Model):
     xpath = models.TextField()
     description = models.CharField(max_length=200)
     rank = models.PositiveSmallIntegerField()
-
-
-class XPathQuerySerializer(serializers.ModelSerializer):
-    class Meta:
-        model = XPathQuery
-        fields = ['id', 'canonical', 'xpath', 'description', 'rank']
-
-
-class MweQuerySerializer(serializers.Serializer):
-    xpath = serializers.CharField()
-    description = serializers.CharField()
-    rank = serializers.IntegerField()
